Added ft_*printf functions, working out of the box for now

+++ src/io/ft_printf/ft_vsprintf.c
@@ -6,13 +6,13 @@
 /*   By: cacharle <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/21 02:34:17 by cacharle          #+#    #+#             */
-/*   Updated: 2019/11/21 03:44:24 by cacharle         ###   ########.fr       */
+/*   Updated: 2020/05/16 13:19:12 by charles          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "libft_printf.h"
 
 int	ft_vsprintf(char *str, const char *format, va_list ap)
 {
-	return (ft_vsnprintf(str, INT_MAX + 1, format, ap));
+	return (ft_vsnprintf(str, (size_t)INT_MAX + (size_t)1, format, ap));
 }
